// backend/src/config/planos.d.ts
export declare const PLANOS: {
    free: {
        nome: string;
        preco: number;
        disparos_mes: number;
        instancias: number;
        leads: number;
        buscas_dia: number;
        templates: number;
        campanhas: number;
        followups: boolean;
        busca_plataforma: boolean;
        enriquecimento: boolean;
        ia: boolean;
        usuarios: number;
        suporte: string;
    };
    starter: {
        nome: string;
        preco: number;
        disparos_mes: number;
        instancias: number;
        leads: number;
        buscas_dia: number;
        templates: number;
        campanhas: number;
        followups: boolean;
        busca_plataforma: boolean;
        enriquecimento: boolean;
        ia: boolean;
        usuarios: number;
        suporte: string;
    };
    pro: {
        nome: string;
        preco: number;
        disparos_mes: number;
        instancias: number;
        leads: number;
        buscas_dia: number;
        templates: number;
        campanhas: number;
        followups: boolean;
        busca_plataforma: boolean;
        enriquecimento: boolean;
        ia: boolean;
        usuarios: number;
        suporte: string;
    };
    enterprise: {
        nome: string;
        preco: number;
        disparos_mes: number;
        instancias: number;
        leads: number;
        buscas_dia: number;
        templates: number;
        campanhas: number;
        followups: boolean;
        busca_plataforma: boolean;
        enriquecimento: boolean;
        ia: boolean;
        usuarios: number;
        suporte: string;
    };
};
export type PlanKey = keyof typeof PLANOS;
//# sourceMappingURL=planos.d.ts.map